refactor: use WeatherContext hook
